Add checks and split code in killvirtv2v.rb

+++ content/automate/ManageIQ/Transformation/TransformationHosts/Common.class/__methods__/killvirtv2v.rb
@@ -9,26 +9,21 @@ module ManageIQ
               @handle = handle
             end
 
-            def main
+            def task_virtv2v_state(task, transformation_host)
               require 'json'
-
-              task = @handle.vmdb(:service_template_transformation_plan_task).find_by(:id => @handle.root['service_template_transformation_plan_task_id'])
-
-              # Retrieve transformation host
-              transformation_host = @handle.vmdb(:host).find_by(:id => task.get_option(:transformation_host_id))
-
-              # Skip if virtv2v has not started yet or is already finished
               return if task.get_option(:virtv2v_started_on).blank? || task.get_option(:virtv2v_finished_on).present?
-              # Skip if virtv2v wrapper information is not available
               return if task.get_option(:virtv2v_wrapper).blank?
-              # Get the state of virtv2v from the conversion host
               result = Transformation::TransformationHosts::Common::Utils.remote_command(task, transformation_host, "cat '#{task.get_option(:virtv2v_wrapper)['state_file']}'")
-              # Skip if no information is available
               return if !result[:success] || result[:stdout].empty?
               virtv2v_state = JSON.parse(result[:stdout])
+            end
+
+            def main
+              task = @handle.vmdb(:service_template_transformation_plan_task).find_by(:id => @handle.root['service_template_transformation_plan_task_id'])
+              transformation_host = @handle.vmdb(:host).find_by(:id => task.get_option(:transformation_host_id))
+              virtv2v_state = task_virtv2v_state(task, transformation_host)
               @handle.log(:info, "VirtV2V State: #{virtv2v_state.inspect}")
-              # Kill virt-v2v process
-              Transformation::TransformationHosts::Common::Utils.remote_command(task, transformation_host, "kill -9 #{virtv2v_state['pid']}")
+              Transformation::TransformationHosts::Common::Utils.remote_command(task, transformation_host, "kill -9 #{virtv2v_state['pid']}") if virtv2v_state.present?
             rescue => e
               @handle.set_state_var(:ae_state_progress, 'message' => e.message)
               raise
